+++ NEWS.md
@@ -1,7 +1,16 @@
+# haldensify 0.0.6
+
+* The core cross-validation routine in `haldensify` for fitting HAL models has
+  been slightly abstracted and moved to the new function `fit_haldensify`.
+* The `haldensify` wrapper function serves to cross-validate over choices of
+  the histogram binning strategy and the number of bins.
+* The defaults of `haldensify` have been changed based on results of simulation
+  experiments; the unnecessary argument `seed_int` has also been removed.
+
 # haldensify 0.0.5
 
 * First CRAN release.
-* Documentation updates
+* Documentation updates.
 
 # haldensify 0.0.3